LoginActivity: Now in Login with google the progress dialog is visible during processing.
DetailsActivity:
Locks the fields when the deleteComment and deleteWine methods are processing.
Fixed the problem with the height of the purchase list dialog.
The purchase dialog icon has been modified.
The size of the Imageview has been decreased and ScaleType has been added to the Imageview.
When deleting a purchase record the list of vintages is reloaded.
Changed the flow of adding a new purchase record to an already registered wine.

MainActivity:
Now, the JobServices message starting or ending before the time will be shown only in the debug version.
JobServices will not start immediately after the first login.

Wine Registration:
Modified to delete the 'Pictures / Camera' folder after a photo has been resized.
The purchase dialog icon has been modified.
Modified the maxLength of the price of a purchase.

SettingsActivity:
Fixed the error that users who logged in with Google could not modify the profile name or the preference of the application's opening filter.

Domain:
Wine2 file deleted
Changed the names of Ids of Wine, WineComplement, Purchase and Comments.

FirebaseBase and DataBaseSyncData:
Functions that were used to retrieve records from an old path in the database and write data to a new path were excluded.

WineSynchronization:
Changed the notification priority and set the TimeoutAfter parameter to remove the notification after the task is completed or failed,  only Android api 26+